requirements.txt
PyYAML==5.4.1
requests==2.26.0
pytz==2021.1
tzlocal==2.1
cryptography==3.4.7
